myProjectLinux/requirements.txt
Django==3.2.5
psycopg2-binary==2.9.1
djangorestframework==3.12.4
django-prometheus==2.1.0
psutil==5.8.0
